requirements.txt
pandas
numpy
nltk
spacy
scikit-learn
tqdm

langchain
langchain-openai
python-dotenv

streamlit
